api-standalone/requirements.txt
flask_socketio
amqpstorm
jsonify
sphinx
psycopg2-binary
SQLAlchemy
pyyaml